requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
paho-mqtt>=1.6.0
requests>=2.31.0
python-dotenv>=1.0.0
pillow>=10.0.0
waitress>=2.1.0
